+++ astropy/io/fits/tests/test_header.py
@@ -1326,6 +1326,10 @@ class TestHeaderFunctions(FitsTestCase):
         assert list(header.keys())[-1] == 'TEST2'
         assert list(header.keys())[-3] == 'TEST1'
 
+    def test_remove(self):
+        # TODO: Test the Header.remove() method; add support for ignore_missing
+        pass
+
     def test_header_comments(self):
         header = fits.Header([('A', 'B', 'C'), ('DEF', 'G', 'H')])
         assert (repr(header.comments) ==